balanced messages separated from rlp lib
