+++ .gitignore
@@ -9,3 +9,4 @@ target
 .DS_Store
 Cargo.lock
 test_snapshots
+.VSCodeCounter
