+++ contracts/balanced_doller/src/balance.rs
