+++ contracts/balanced_doller/src/balanced_dollar.rs
@@ -11,8 +11,8 @@ use crate::{
     config::{get_config, set_config, ConfigData},
     xcall_manager_interface::XcallManagerClient,
 };
-use soroban_rlp::address_utils::{get_address_from, is_valid_bytes_address};
-use soroban_rlp::messages::{
+use soroban_rlp::balanced::address_utils::{get_address_from, is_valid_bytes_address};
+use soroban_rlp::balanced::messages::{
     cross_transfer::CrossTransfer, cross_transfer_revert::CrossTransferRevert,
 };
 use soroban_token_sdk::TokenUtils;
