+++ contracts/balanced_doller/src/contract.rs
