+++ contracts/balanced_doller/src/errors.rs
