+++ contracts/balanced_doller/src/metadata.rs
