+++ contracts/balanced_doller/src/states.rs
