+++ contracts/balanced_doller/src/storage_types.rs
