+++ contracts/balanced_doller/src/xcall_manager_interface.rs
