+++ libs/soroban-rlp/src/balanced/messages/cross_transfer.rs
