+++ libs/soroban-rlp/src/balanced/messages/mod.rs
@@ -4,5 +4,4 @@ pub mod withdraw_to;
 pub mod cross_transfer;
 pub mod cross_transfer_revert;
 pub mod configure_protocols;
-pub mod execute;
