+++ libs/soroban-rlp/src/lib.rs
@@ -1,6 +1,5 @@
 #![no_std]
-pub mod address_utils;
 pub mod decoder;
 pub mod encoder;
-pub mod messages;
 mod utils;
+pub mod balanced;
